+++ CHANGES.md
@@ -3,6 +3,12 @@
 LogTape changelog
 =================
 
+Version 0.11.1
+--------------
+
+To be released.
+
+
 Version 0.11.0
 --------------